Architecture on sheet building, sheet style and trimming.
Trimmed strings that contained common, unnused, values on report descriptions.
Added custom styles to sheet headers and cells.
Unified all the sheet building classes into one class.
Added month reference to the name of the output file.
Added some methods on the utils for legibility and easy boolean checking.
Changed sheet compilation window. Now, the search will begin from the first salary found, and end on the next salary, or the end of the table. The compilation month will be the month of the first salary.

TO DO: Write tests, pmd and setup code to run for more than one month at a time.