Corrected a bug with dates on built sheet, and implemented branched expenses.
It is now possible to define an expense as [branchedExpenses];[originalBranch]. For example, if a bank report is declared as "50;f", it means that the report belongs to the fixed expenses, but R$ 50,00 belongs to the expense counterpart (mom expenses). The value that goes to fixed expenses is subtracted by 50, and a report of -50 is added to mom expenses. This way, a bank report can be shared between two entities.